Add a very simple Engine
This commit adds a very simple engine implementation that can be used
to generate an execution context based on a WITX definition, and run
WebAssembly components that implement the WITX.

The implementation uses Wasmtime and `witx-bindgen`, and the test is
an echo implementation written in Rust.

Currently, the implementation does not test additional features, such as
custom linker imports.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>