feat: first pass at implementing a better engine
This commit starts adding an actual implementation for the new execution
engine. Specifically, it parses the new application configuration,
pre-instantiates the modules required for the application, and returns
the execution context so it can be used by one, or potentially multiple,
higher-level object.

The two important aspects missing from this commit are:

* parsing the environment and files required for configuring WASI
* allowing outbound HTTP requests, because of some current limitations
in both implementations.

Future commits should enable both scenarios.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>