feat: re-implement HTTP engine
This commit updates the implementation of the HTTP engine to work with
the new configuration and execution context.

This introduces an HTTP trigger that takes an application configuration
(with local, "core" components, meaning it assumes they have been
pulled and linked in a previous step), constructs a route map with the
routes handled by each component, and starts the HTTP listener.

Future commits might update this to take a *list* of application
configurations, as opposed to a single application.

This commit also introduces the concept of an HTTP executor. This is a
specific HTTP implementation, which will allow us to properly version
Spin HTTP interfaces and potentially introduce a Wagi executor for
supporting languages that don't currently have support for Wasm
components.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>