chore: temporarily remove Hippo implementation
The early prototype of Spin included very basic support for creating and
deploying an application to Hippo.

As outlined in https://github.com/fermyon/spin/issues/22, we want to have
much better support for integrating with Hippo, which we will get from
using the auto-generated OpenAPI Rust client,
https://github.com/fermyon/hippo-openapi, potentially through consuming
https://github.com/deislabs/hippo-cli directly.

As we figure out how to best integrate (subcommands, plugins), there is
no point in having broken code that will not be supported in the future.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>