chore: remove Cargo.toml indentation
This commit reverts the (implicit) decision to indent all TOML files and
align entries to the right.

It was a personal preference that is not in line with what the community
uses for TOML files.

I still have a strong preference towards alphabetically sorted dependencies
in the Cargo.toml file.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>