feat: strip component route from request URI
This commit updates the request URI passed into an HTTP component from
the full URI from the server to a URI relative to the component's
declared handled route.

For example, if `route = "/foo/..."`, and a request comes for `/foo/bar`,
the URI received by the component is now `/bar`. This also handles
wildcard route patterns.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>