chore: refactor the assets module
This commit refactors the assets module from the execution context
crate with the main purpose of having a consistent style between the
modules of the same crate.

Note that this does not change any of the logic in the module at all.
Specifically, the commit does a few main things:

    * it renames most functions and identifiers to be slightly less
    verbose, and adds comments where needed. This, together with
    running `rustfmt` again on this module, makes the style similar to
    the rest of the project.
    * it updates most functions that took two separate `impl AsRef<Path>`
    parameters to have generic `AsRef<Path>` arguments.
    * simplifies the use of Rust maps and removes unnecessary `let`
    bindings where possible.

Finally, the one change in this commit that has an impact on the
functionality is marking all copied files as read-only.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>