chore: refactor the application loader
This commit starts the refactoring of the way a Spin manifest file is
loaded from various sources into a configuration runnable by the
execution context.

Specifically, it modifies the `config` crate to only contain the
common configuration, then adds a `loader` crate that constructs
the core configuration from the source, including preparing the static
assets.
This means the execution context is no longer responsible for copying
assets, and receives a list of directory mounts for each component.

A future commit will update the `loader` crate with functionality
to pull from Bindle.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>