feat: run applications from Bindle
This commit re-adds the ability to run applications given a bindle ID
and server URL.
It also adds the option to pass a byte array as a component module
source, thus avoiding unnecessary IO by copying files.
This is currently used in the Bindle loader, but it could be used
in the local load as well in the future.

Co-authored-by: Ivan Towlson <ivan.towlson@fermyon.com>

Signed-off-by: Radu Matei <radu.matei@fermyon.com>