feat: allow passing environment variables from CLI
This commit enables adding environment variables through CLI options
to all components of an application when starting with `spin up`.

This implementation mutates the component configuration, and would have
to be replicated for all future triggers.
A more generic approach might make more sense (the `prepare_component`
function in the execution context allows taking an optional map of
environment variables, but this is the smallest change that gives us the
ability to override environment variables on the CLI).

Signed-off-by: Radu Matei <radu.matei@fermyon.com>