fix(tests): fix wrong timeout in integration test
This commit fixes a wrong timeout in the integration test logic that
added a 120-second sleep cycle for tests when a particular server
(Bindle or Spin) wasn't ready, as opposed to waiting for one second,
for a total of 120 times when needed.

This commit also temporarily removes the unused release CI workflow
altogether.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>