ref(crates/engine): sanitize log file paths
This commit adds logic that sanitizes the log file path.
Two log files are created for every component (one for
stderr, one for stdout) and the default path for these
files is generated using the app name and the component id.
Both the app name and the component id are sanitized before
being added to any paths. This helps us bypass issues related
to a user inputing app names and component ids which may contain
unsafe characters or characters that will have unintended
consequences if used as part of a path.

resolves #163

Signed-off-by: Michelle Dhanani <michelle@fermyon.com>