feat(deploy): add spin deploy command
This is an MVP for the spin deploy command. Given
a bindle server url, hippo server url, and hippo
username/password, this command will create
(1) and push a bindle containing the app to the bindle server
(2) a Hippo App with the name provided in the Spin.toml
(3) a Hippo channel in the Hippo App called "spin-deploy"
and create a post request to register a revision in Hippo.
If an app with the same name already exists, this command will
return an error. If a channel with the name spin-deploy exists,
this will return an error. These errors will be handled more
gracefully with prompts/helpful error messages in subsequent PRs.

This commit it meant to lay the groundwork for the deploy command
so folks can start playing with it and collect opinions.