chore(ci): stop cancelling concurrent runs
https://github.com/fermyon/spin/pull/170 introduced cancelling in-progress
workflow runs for builds, which made sense when we were constrained on build
minutes.

However, this seems to cancel in-progress runs quite often for both in-flight
PRs *and* for builds on `main`, which means that we regularly have to manually
re-run builds for both PRs and on main (when the overall status of the repository
is failed, because a job was cancelled).

This commit removes the "cancel-in-progress" condition for builds.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>