feat(cli): add Wasmtime cache
This updates the `spin up` command to always use a Wasmtime cache.
Specifically, if no option is passed, it will attempt to use the default
Wasmtime cache, or create a new `enabled` cache configuration on the first
run.

This behavior can be overridden with the `--disable-cache=true` setting, and the
cache file can be configured with the `--cache <path-to-cache-file>` flag.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>