chore(rustfmt): standardize the order of Rust imports
This commit updates the order of the Rust imports throughtout the project to be
consistent with — specifically, standard library, external, and crate imports.

The `group_imports=StdExternalCrate` rustfmt setting only works for the nightly
toolchain, so this commit does not enforce it in CI (which would also require
all contributors to run the nightly toolchain to format).

To run the formatter using this setting:

`cargo +nightly fmt --all -- --emit files --config group_imports=StdExternalCrate`

Signed-off-by: Radu Matei <radu.matei@fermyon.com>