doc(quickstart): specify tag in `git clone` command
Per #615, the HEAD of the main branch of the repo is not guaranteed to be
compatible with the most recent SDK release, so it's better to have users use
the latest release tag.

We also need better error reporting for SDK compatibility issues; I'll open a
new work item to track that.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>