fix: Terminate trigger handler on `spin up` termination
The `Ctrl-C` handler works because it is handled in the trigger handler
child process, but if the parent `spin up` process is terminated
instead - as with e.g. Nomad - the child is left running.

Fix this by adding a signal handler to the parent process which sends
SIGTERM to the child.

Signed-off-by: Lann Martin <lann.martin@fermyon.com>