Add `spin-conformance` crate (#724)
This crate provides a suite of tests to determine which features of Spin a guest
module is capable of exercising.  The suite may be run via the `test` function,
which accepts a `wasmtime::Module` and returns a `Report` which details which
features the module does or does not support.  Alternatively, there is also a
CLI version which generates a report in JSON form.

See the doc comments for `Report` and friends for details on what is tested.

Note that the current implementation is quite simple, and there are a few ways a
module implementation could "cheat", i.e. hard-code output to match what the
test expects without actually doing the thing it was asked to do.  We could
prevent that by randomly generating parameters, but I don't consider that a high
priority.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>