deploy: Shouldn't complete until new version is active
Before this change, a redeploy of an existing app would be seen as
"ready" immediately, since the previous version was ready. This meant
that after a successful deploy a user might still see the old version
active for a while.

This checks the new "app info" endpoint in the readiness check loop, and
only completes if there isn't an old version active.

Signed-off-by: Lann Martin <lann.martin@fermyon.com>