chore(GitHub): Refactor build/test workflow
Our current builds are currently unstable because we run out of disk
space on the default GitHub Actions builders.

While we could start trying to be clever and purge various resources
between steps - it's simpler to refactor the build into mutliple
independent steps.

For now we're going to use a "shared-for-the-platform" cache based on
the hash of the Cargo.lock. I plan to follow up with a little more
rigerous of a cache-key system that can also naturally expire a little
sooner to ensure everything is still buildable - but for now, lets get
spin building consistently again.

Signed-off-by: Danielle Lancashire <dani@builds.terrible.systems>