add `--direct-mounts` option to `spin up`
This option tells `spin up` to directly mount any asset directories specified in
the manifest file instead of copying the contents to a temporary directory and
mounting that.  This is useful when iteratively developing the static assets in
the app, since any modifications to assets are immediately visible to subsequent
invocations -- no restart required.

Closes #964.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>