Add initial cache for Spin URL component sources
Before this commit, for a `spin.toml` file that had a URL module source
for a component (such as the file server and redirect templates),
subsequent runs of `spin up` would always download the module from the
URL before starting the application.

https://github.com/fermyon/spin/pull/1014 added a content-addressed
registry cache that can be used to store remote module sources.

This commit updates the Spin local loader to use that cache to store
the component sources referenced through URL in `spin.toml`.

Before fetching a component, the loader will check the local cache
first. If present, it will read the file and return its content;
otherwise, it will fetch it, store it in the cache, then return its
content.

Currently, the default cache root is always used. Once `spin up` will
have a way to configure it, that will be propagated here as well.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>