Rename `spin oci` to `spin registry`
As I have started using the OCI functionality a bit more, I realized
that "OCI" is a fairly specialized three letter acronym, and expecting
a user to know it and immediately understand "this has something to do
with registries" might not be reasonable.

This commit changes the top level `spin oci` command to `spin registry`,
while still keeping an alias for `spin oci`.

This should be more intuitive for a user.
Happy to close this PR if people think `spin oci` is more specific.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>