ref(build): updates build to allow interactive commands
Previously, spin build would hang for commands that were
interactive (like a login command). This was a result of
how we were setting up standard streams for the subprocess
created to run the build command. This change updates how
we set up standard streams such that the subprocess
inherits standard streams from the parent which  enables
interactive commands.

resolves #1026

Signed-off-by: Michelle Dhanani <michelle@fermyon.com>