split `redis-types::value` into two separate types
The `nil` case doesn't make sense for parameters, and string arguments
have different interpretations than Redis status results, so they
should have different name.

This also renames cases to be consistent with the names in the
database WIT files and with the semantics of Redis status messages.

The name `redis-result` is admittedly somewhat verbose, but I wanted
to avoid clashing with Rust's `Result` type and also the HTTP
`Response` type, which didn't leave me with any natural, succinct
names.  And, to be consistent, I went with `redis-parameter` for the
other type.  Naming is hard.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>