add general-purpose `execute` function to `outbound-redis` (#1134)
* add general-purpose `execute` function to `outbound-redis`

By popular request, this function should allow executing any arbitrary
Redis command and retrieving its result, if any.  Ideally, we'd
support a statically typed interface for every command, but that would
be a pretty huge effort, so we settle for dynamic typing for now.
Once we have TCP socket support, users will be able to use their
favorite Redis client library and enjoy static typing at that level,
if applicable.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* fix Windows CI warning

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* Update tests/outbound-redis/http-rust-outbound-redis/src/lib.rs

Co-authored-by: itowlson <github@hestia.cc>
Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* split `redis-types::value` into two separate types

The `nil` case doesn't make sense for parameters, and string arguments
have different interpretations than Redis status results, so they
should have different name.

This also renames cases to be consistent with the names in the
database WIT files and with the semantics of Redis status messages.

The name `redis-result` is admittedly somewhat verbose, but I wanted
to avoid clashing with Rust's `Result` type and also the HTTP
`Response` type, which didn't leave me with any natural, succinct
names.  And, to be consistent, I went with `redis-parameter` for the
other type.  Naming is hard.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* remove `redis-parameter::str` in favor of `binary`

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

---------

Signed-off-by: Joel Dice <joel.dice@fermyon.com>
Co-authored-by: itowlson <github@hestia.cc>