Implement `spin registry login`
This commit implements functionality for logging in to an OCI registry.
Until this commit, a user would have had to use external tooling to log
in to a target registry (such as `docker login` or other registry
specific tools). This had the advantage that users who already had a
setup for working with container regsitries could continue to use it,
but it introduced a dependency on external tools for pushing a Spin
application.

This commit implements the functionality to login to a registry natively
into Spin.
Specifically, it adds this a `spin regsitry login` command:

```
$ spin registry login --help
Log in to a registry

USAGE:
    spin registry login [OPTIONS] <SERVER>

ARGS:
    <SERVER>

OPTIONS:
    -h, --help                   Print help information
    -p, --password <PASSWORD>    Password for the container registry
        --password-stdin         Take the password from stdin
    -u, --username <USERNAME>    Username for the container registry
```

This command mirrors existing tools that perform login operations to
registries (flags for the username and passwords, flag to read the
password from standard in that takes precedence).

The credentials are base64 encoded (standard for container tools that do
not use Docker's credential helpers) and written to
`$XDG_CONFIG_HOME/fermyon/registry-auth.json` (side-by-side with the
Fermyon Cloud authentication file).

Functionality that needs to work with a registry (`spin registry
push/pull`, `spin up`) will first search the Spin configuration for the
credentials, then will fallback to searching the authenticated
registries in the Docker configuration file.

Examples:

```
$ echo $GHCR_PAT | spin registry login --username radu-matei ghcr.io --password-stdin
Successfully authenticated as radu-matei to registry ghcr.io

$ spin registry login --username radumatei --password <password> https://index.docker.io
Successfully authenticated as radumatei to registry https://index.docker.io

$ spin registry login index.docker.io
Username:
radumatei
Password: (prompt, hidden input)

Successfully authenticated as radumatei to registry https://index.docker.io
```

This functionality should make `spin registry login` a drop-in
replacement for most of the existing tools that login to registries,
while continuing to reuse existing credentials as a fallback.

A future commit needs to make writing the configuration file more
flexible, and add unit tests for decoding credentials.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>