lazily create and reuse store manager for each app
My previous refactor introduced a performance and logging verbosity
regression such that we created a separate store manager for each
component of each app on each request, which is a lot more expensive
than doing it once per app and reusing it across requests.  It also
resulted in printing the database location on every request, which was
annoying.

Now we lazily create and reuse the store manager, printing the message
only once when the database is first opened.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>