export SDK version and language functions per SIP 011
Per
https://github.com/fermyon/spin/blob/main/docs/content/sips/011-component-versioning.md,
this commit adds `spin-sdk-version-$VERSION`,
`spin-sdk-language-$LANGUAGE`, and `spin-sdk-commit-$COMMIT` exports
to any module using the Rust or Go SDKs.

Regarding the version, we'll want to update it before and after every
Spin release _and_ each time the ABI or API changes.  See the SIP for
details.

Note that the `spin-sdk-language-$LANGUAGE` export is gated on a
default-enabled feature for the Rust SDK, since the JS and Python SDKs
will want to override it.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>