feat(sdk/rust): add optional feature for serde
This commit updates the Rust SDK to include two new functions that
integrates key/value with serde. Specifically:

- `set_json`: this function takes a key and a struct that can be
  serialized by serde into JSON, and that representation is added to the
  store
- `get_json`: this function takes a key, reads the key/value pair from
  the store, then deserializes it into the specified type

Considering that storing serialized state in Spin's key/value store is
one of the popular use cases for the API, these convenience functions
should simplify that operation.

This functionality is added as an optional feature in the SDK, so it
doesn't add to the binary size of the resulting component unless needed.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>