Enable the pooling allocator by default
This commit enables Wasmtime's pooling instance allocator by default.
The pooling allocator performs better than the default allocator for
many concurrent short-running instances which Spin is tailor-made for.

The intention behind this commit is that everyone can enable the pooling
allocator without issue. If that doesn't turn out to be the case this
can be switched to an explicit opt-in.

Additionally the hope is that there's no need to configure the pooling
allocator. Currently there are no options for it although there are some
environment variable escape valves if necessary to avoid too-bad
breakage. The parameters here for the pooling allocator primarily impact
the virtual memory reservation made by the pooling allocator and in
theory are good enough for all current-day reasonable wasm modules.

Signed-off-by: Alex Crichton <alex@alexcrichton.com>