Update HTTP benchmarks
This commit is an attempt at enhancing the HTTP benchmarks in-repo with
some more flavors in addition to changing what's being measured. One of
the important vectors measured previously was the effect of concurrent
requests on latency, but the measurement reported was one of handling an
entire batch of requests. Instead this commit now reimplements this in
terms of what the latency of a single request is with N units of
background work.

Benchmarks will now, before running, spin up background work performing
concurrent requests. The benchmark then executes, measuring the latency
of a single request while there's background work. When the measurement
is over all background requests are torn down.

Additionally this commit fixes the processing of the `sleep` parameter
in the `spin-http-benchmark` crate since the `params` field it
previously read is no longer set.

Signed-off-by: Alex Crichton <alex@alexcrichton.com>