add `wasi-http` trigger support based on `wasmtime-wasi-http`
This required upgrading to a snapshot of what will eventually be Wasmtime 14, as
well as upgrading `wit-bindgen` to 0.12.

Note that I had to make a copy of wit/preview2 under sdk/rust/macro in order to
allow the `http_component` and `redis_component` macros to generate a
`wit_bindgen::generate` invocation that actually works.  Ideally we'd avoid this
duplication, e.g. by using a registry.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

address PR feedback

- Move `spin_http::empty` and `spin_http::full` into a `body` submodule
- Use `Bytes::from_static` where appropriate

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

update Wasmtime and `spin-componentize` deps to Wasmtime release-14.0.0 branch

This also re-enabled wit-bindgen 0.8 support in `spin-componentize` to
(hopefully) fix the e2e tests.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

add `wasi_http_component` macro to Rust SDK

The implementation is pretty ugly given that the inbound and outbound parts of
`wasi-http` are tightly coupled, and it's not currently possible to tell
`wit_bindgen::generate` to reuse types instead of generating duplicate ones.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>
Signed-off-by: Ryan Levick <ryan.levick@fermyon.com>