add `wasi-http` trigger support based on `wasmtime-wasi-http`
This required upgrading to a snapshot of what will eventually be Wasmtime 14, as
well as upgrading `wit-bindgen` to 0.12.

Note that I had to make a copy of wit/preview2 under sdk/rust/macro in order to
allow the `http_component` and `redis_component` macros to generate a
`wit_bindgen::generate` invocation that actually works.  Ideally we'd avoid this
duplication, e.g. by using a registry.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>