up: Split app resolution into two stages
First we "resolve" the AppSource into a ResolvedAppSource. This step
isn't expected to "load" an app fully, e.g. copying static assets.

At this stage we have enough information to determine the trigger
executor, which lets us serve `--help`.

Then we "load" the ResolvedAppSource into a LockedApp as before.

Note that this change is basically a no-op for OCI apps as the changes
to split these stages are incompatible with the rest of the OCI code
today.

Signed-off-by: Lann Martin <lann.martin@fermyon.com>