rust/sdk: make exporting symbols opt-out
Without it, there are linker errors when both a driver crate
(e.g. libsql-client-rs) and a user app have spin-sdk as their
dependency. That's because both contain global symbols (T in nm output)
and the linker is not able to determine which one is to stay.

Before the patch:
```
[sarna@sarna-pc rust]$ cargo build
[sarna@sarna-pc rust]$ nm -a ../../target/debug/libspin_sdk.rlib | grep T\ spin-sdk-
0000000000000000 T spin-sdk-commit-8d4334eca6eb2b9cec0105612cc9a38402f839b4
0000000000000000 T spin-sdk-language-rust
0000000000000000 T spin-sdk-version-1-5
```

After:
```
[sarna@sarna-pc rust]$ cargo build --no-default-features -F http,json
[sarna@sarna-pc rust]$ nm -a ../../target/debug/libspin_sdk.rlib | grep T\ spin-sdk-
nm: lib.rmeta: no symbols
```