Always url encode authority when parsing OutboundUrl
This prevents errors when users don't url encode the authority where
the authority accidentally gets viewed as the url itself.

For example, in `mysql://user:pass#word@xyz.com` the `user` will be parsed
as the host before this change. With this change the authority is
url encoded and correctly parsed as the authority.

Signed-off-by: Ryan Levick <ryan.levick@fermyon.com>