add support for `wasi:http/incoming-handler@0.2.0-rc-2023-11-10`
This builds upon Alex's excellent work in
https://github.com/fermyon/spin/pull/2108, adding the ability to run guests
which target the `0.2.0-rc-2023-11-10` version of `wasi-http`.

I've added a test which uses `wit-bindgen` and `wit-component` directly to
ensure it uses the correct snapshot for everything `wasi-http`, `wasi-cli`, etc.

Note that we're not yet updating any SDKs to use the new snapshot, since we
still want people to be able to use the new SDK to target Spin 2.0.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>