update to Wasmtime 17 and WASI 0.2.0-rc-2023-12-05
As of this writing, Wasmtime 17 has not yet been released, so we're temporarily
pointing at the `release-17.0.0` branch.  By the time it is released, WASI
0.2.0-rc-2023-12-05 will become WASI 0.2.0, at which point we'll need to update
the WIT files and code to match.

The upshot is that we'll be supporting three snapshots indefinitely: 2023-10-18,
2023-11-10, and 2023-12-05 AKA the final 0.2.0 release.

Most of the changes here deal with supporting those three snapshots, while the
rest deal with the Hyper 1.0 update, which led to a bit of awkwardness when
mixing Hyper and Reqwest, since the latter has not yet been updated to use Hyper
1.0.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>