point Go templates to `v2.2.0` instead of `main`
CI used to take care of this when the SDKs lived in this repo and were versioned
in lock-step with Spin itself.  Now that neither of those are true, we'll need
to update the versions manually as necessary.

Once this is merged into `main`, I'll backport it to the 2.3.0 tags.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>