fix(oci/config): ensure unique OCI image config
This commit ensures that applications pushed to OCI have unique image
config fields for unique Spin application content and metadata by adding
a label in the OCI image config to the content digest (SHA256) of the
Spin locked application file.

This is to address the issue of the Containerd Spin shim serving
outdated content, because all images of Spin apps on a node would have the
same image ID (the content digest of the OCI config object, which was
identical for all Spin apps).

ref https://github.com/spinkube/spin-operator/issues/40

Signed-off-by: Radu Matei <radu@fermyon.com>

Co-authored-by: Rajat Jindal <rajatjindal83@gmail.com>
Co-authored-by: Danielle Lancashire <dani@builds.terrible.systems>
Co-authored-by: Michelle Dhanani <michelle@fermyon.com>