Refactor `TriggerExecutor` to make `InstancePre` an associated type
Moves the `EitherInstance*` logic into just the `spin-trigger-http`
crate where the redis trigger, for example, only has to deal with
components. The intention of this is to open up future customization of
instances within a trigger that don't necessarily need to affect all
other triggers.

The new associated type is bound itself by a new trait
`TriggerInstancePre` which encapsulates the required functionality. This
trait has a default implementation for `spin_core::InstancePre<T>` which
serves for triggers that only work with components.

Signed-off-by: Alex Crichton <alex@alexcrichton.com>