Lift http handler type discovery up a layer
This commit lifts the discovery of what type of handler a component is,
for example spin-vs-wasi-vs-wasi-snapshot, up one layer from where it
currently resides. Previously this determination was made on each
request and now it's made when a component first loaded and
pre-instantiated instead. This would shift possible errors earlier in
the system, for example.

This removes a bit of per-request work and makes the creation of the
handler in `execute_wasi` a bit more straightforward with less guessing
of exports for something that was already probed.

Signed-off-by: Alex Crichton <alex@alexcrichton.com>