Update Wasmtime to 20.0.2
This commit updates the Wasmtime version used by Spin from 18.0.4 to
20.0.2. No motivation beyond keeping things up-to-date. This involved a
number of minor WASI refactorings such as moving some items around and
some refactorings around the `wasmtime-wasi-http` traits-and-types.
Notably the chained handler was switched from `anyhow::Result` to
`wasmtime_wasi_http::HttpResult` which helps make more explicit which
errors are traps and which are normal error codes.

Also update the MSRV of Spin to 1.75 as that's what Wasmtime requires

Signed-off-by: Alex Crichton <alex@alexcrichton.com>