Fix warning from redis crate.
You can read about the warning here: https://github.com/redis-rs/redis-rs/issues/1228

Essentially, the functions in question are subject to a subtle change
that will be coming in Rust 2024. To get ahead of this change, a warning
is now produced in Rust 1.81. The fix is small even if it's
unfortunately a bit awkward.

Signed-off-by: Ryan Levick <ryan.levick@fermyon.com>