Update Wasmtime to 29.0.1
This commit updates Wasmtime from 25.0.3 to 29.0.1, the latest release
of Wasmtime. This update is accompanied with few minor changes around
bindings generation such as traits no longer needing `#[async_trait]`.

Otherwise one of the main features is that the WITs included with
Wasmtime's WASI implementation have been updated to 0.2.3 as well. This
should continue to still work with guests using 0.2.2 and prior, though.

Signed-off-by: Alex Crichton <alex@alexcrichton.com>