Update wit-bindgen used in tests
This commit updates the `wit-bindgen` crate version used in tests from
0.13.0 and 0.16.0 to 0.32.0. A significant number of changes have
happened to `wit-bindgen` in the intervening releases which relatively
radically change how, for example, exported functionality works. The
goal here was to get all the tests compiling/passing again with the
updated version of wit-bindgen, but this otherwise doesn't refactor
things too much.

Note that 0.32.0 isn't the latest `wit-bindgen` because the latest
version needs other changes in WIT files to replace usage of `float32`
with `f32` for example. The full update here will be deferred to #3005

Signed-off-by: Alex Crichton <alex@alexcrichton.com>